refactor: split up loader
